NPC edit adjustments

+++ zone/zonedump.h
@@ -36,8 +36,8 @@ struct NPCType
 {
 	char	name[64];
 	char	lastname[70];
-	int32	current_hp;
-	int32	max_hp;
+	int64	current_hp;
+	int64	max_hp;
 	float	size;
 	float	runspeed;
 	uint8	gender;
